Impl `ValidationCheck` for `CurveGeometryMismatch`
